Create reusable Subject Selector Component
